Make camel-k compatible with camel-3 M2
