+++ camel-k-runtime-groovy/src/test/groovy/org/apache/camel/k/groovy/LoaderTest.groovy
@@ -19,6 +19,7 @@ package org.apache.camel.k.groovy
 import org.apache.camel.impl.DefaultCamelContext
 import org.apache.camel.k.InMemoryRegistry
 import org.apache.camel.k.Source
+import org.apache.camel.k.adapter.Routes
 import org.apache.camel.k.support.RuntimeSupport
 import org.apache.camel.model.ToDefinition
 import spock.lang.Specification
@@ -44,7 +45,8 @@ class LoaderTest extends Specification {
             def routes = builder.routeCollection.routes
 
             routes.size() == 1
-            routes[0].inputs[0].endpointUri == 'timer:tick'
             routes[0].outputs[0] instanceof ToDefinition
+
+            Routes.getInput(routes[0]).endpointUri == 'timer:tick'
     }
 }
